Refactor events controller to move instagram endpoints to another controller.
